7766 FIX Raw Edition: Fix missing default ruleset and automation user
When creating a new site with Checkmk Raw Edition 1.6.0b1 the sample
configuration could not be created correctly. It was missing the empty
sample Event Console rule pack and the builtin "automation" user.

In the moment this happens an error like this is visible in var/log/web.log:

C+:
UnboundLocalError: local variable 'cmk' referenced before assignment
C-:

Change-Id: I5c3c48f6f0404c0e6ea4564090caa20834a281a5