Modularied WATO sample config creation
The _create_sample_config() in watolib.__init__ imported several modules
in a way that it was not allowed to. Because these imports were
resulting in circular dependencies the imports were done locally in the
function. This could lead to errors e.g. when creating the sample config
for Raw Edition installations.

The registration of sample configurations has now been moved to a central
class registry object in cmk.gui.plugins.wato.utils which can be
imported from all necessary places.

The watolib._create_sample_config function is then calling all registered
sample config generators in a defined order.

One difference is that the sample configuration executes all generators
in the desired order, even when a single generator fails. In such a
situation an error is written to the web.log.

Change-Id: I1cb412f4373ebfd3027dbf0b538dbab6dc85bb96