7775 FIX Fix some NagVis maps snapin tree rendering issues
The NagVis maps snapin can be configured to be rendered as tree
(by configuring the global setting "multisite_snapin_layout" in NagVis).

The fold arrows rendered in the NagVis snapin were not displayed correctly
in 1.6.0b1. The rendering has now completely been cleaned up to be working
as expected again.

The folding state was never remembered for this snapin which has also been
fixed now.

Change-Id: Iee5a633bca6e612be1458c1eea35a22489dc5910