7421 ucs_bladecenter_if, winperf_if: discover single interfaces of interface groups
Individual interfaces for portchannels of ucs_bladecenter_if
and teamed interfaces of winperf_if are now discovered additionally
to the grouped interfaces. This makes troubleshooting for grouped
interfaces easier since the metrics of the individual interfaces
are present as well. The new services of the individual interfaces
have to be added to the monitoring.

CMK-1473

Change-Id: Id8416816cdaf30c4f4c327c66649b156ed6a2081