if.include: always use the correct timestamp in check_if_common
The timestamp used in check_if_common was only used for the
get_rate calls, but not passed to the subsequent calls of
check_if_common_single. Therefore check_if_common_single
calculated an own timestamp even if a timestamp was already
given. This lead to a wrong reference time in winperf_if and
therfore to wrong calculated values.

CMK-1473

Change-Id: I7d70a434659e8b8a23b9bba01bd6fcdb295994d6