7168 FIX cmk-update-agent.exe: Fix erroneous update validation
Regression since 1.5.0p17 and Werk #7167:

The installation validation introduced in Werk #7167 fails to read
the agent hash correctly from a newly performed update.

As a result, the update gets invalidated, which you will observe within
the "Check_MK Agent" check that shows

<pre> WARN - error: MSI Installation failed. </pre>

Contrary to this warning, the msi is not broken.

With this Werk, the correct behavior gets restored.
However, due to inconsitencies between the saved update state and the
actually installed Windows Agent, to restore all to normal behavior,
it might be necessary to update the agent (bake new agents for the affected
windows hosts, wait for an automatic update or trigger it manually) twice.

(CMK-2226)

Change-Id: Iaf82dd2d8e4dcd812ba0067002d3ce7cb2e8e3cf