7424 FIX mk_postgres: add missing database parameter
The mk_postgres plugin was missing a database parameter in
the queries for the checks postgres_bloat and postgres_stats.
Therefore, the name of the database user was used as a
fallback. As a result no output for postgres_stats and
postgres_bloat was returned if a database with the username
did not exist. If a database with the username exists
the checks give the correct output.

To resolve the issue the fixed version of the mk_postgres
plugin has to be deployed on the affected hosts.

Change-Id: I1df8e3e52184c86dff7e2335e29639079fd69bc1