7425 FIX mk_postgres, postgres_instances: fix plugin output for older operating systems
mk_postgres used the pgrep command to find the running
PostgreSQL instances. Unfortunately, version 1.5.0p3
added a command line argument that is not supported on
older operating systems like e.g. SLES 11, but only on
newer ones like e.g. SLES 12. As a result Checkmk did
not find any PostgreSQL instances on older operating
systems.

To resolve the issue, the fixed version of the mk_postgres
plugin has to be deployed on the affected hosts.

Change-Id: I86ea41f2b03b98d9fd33373fa4fbe7263304521f