Filter for Historic Metrics table
Dynamically generated RRD columns are recognized by the filter system early
in the process when they are collected for the view.

This filter is not present in the service datasource despite the filter
specification claiming that "host" enables it in the hosts and services
views.

Filters get always activated, instead of appearing when they are
configured. This causes for example that the SW/HW query always happens,
because that filter is active all the time.

CMK-2182

Change-Id: I79d470334ddaadf16d8aceb10296b7f64a2e37c7