Columns that deal with history data require a Title
Since the content is dynamically generated based on RRD query and then
statistical processing. It is mandatory to label this dynamically generated
columns. A bit of a Hack is used as painter title should be static
properties, nevertheless title functions are allowed for SLA and thus here
we abuse the inheritance from painter and let title be a function with
optional parameters.

Works for GUI and PDF reports

CMK-2185

Change-Id: Icc8579210345fcfd0c91cddb353beafd47dcde6f