7420 FIX ps: only show processes which do not belong to a docker container
On hosts which run Docker the ps section contained processes which belong
to Docker containers. This Werk omits these container processes and only
shows processes which do not belong to a Docker container. If the agent
is run in the context of a Docker container this filter is not applied.

If you want to use this feature a check_mk_agent including this Werk has
to be installed on the host.

FEED-2973

Fixes #4

Change-Id: I81e7e9fd6a861eb82271f7f2746074eaced6732b