7442 FIX Pagerduty notification host identifier to hostname from IP
Pagerduty requires a host identifier to assign an event to the troubled
host. This was the IP address of the host, but in some cases like cloud
Services, there are no IP addresses assigned to a host. Thus the hostname
declared in Checkmk will be used when no IP address is found to declare
which host is under trouble

FEED-3679

Change-Id: Ia12025ee2395fa31d97d8d49f3dd871b9dd304a3