7443 FIX SW/HW inventory filters only query for information when configured
Currently all Checkmk filters are enabled by default although when
unconfigured they remain passive. Most filters process data directly from
the livestatus query but inventory filters perform a new query to obtain
the SW/HW inventory of the host. Now inventory filters only query for data
to filter for when they are actively configured, instead of gathering data
at first and then deciding how to filter.

This fix increases the performance of the views, as fewer queries are done
it thus speeds up the time to show a view.

Change-Id: Ic0f98932832e098a4e7ef01069025729325f9eae