7427 FIX mail: don't escape state markers in emails
Since version 1.5.0p14 state markers in HTML emails were escaped
and not interpreted as HTML but shown as text. The state marker
for CRIT was e.g. shown as:

&lt;b class="stmarkCRITICAL"&gt;CRIT&lt;/b&gt;

This werk restores the old behaviour and state markers are shown
in HTML emails as usual.

If you use the macros HOSTOUTPUT_HTML, SERVICEOUTPUT_HTML, or
LONGSERVICEOUTPUT_HTML in custom notification scripts please
switch to the macros HOSTOUTPUT, SERVICEOUTPUT, and
LONGSERVICEOUTPUT which contain the textual state markers
'(!)' and '(!!)' for WARN and CRIT, respectively. The *_HTML
variables will be deprecated in 1.6 and removed in 1.7.
If you want to replace the textual state markers by HTML code
you can use the utility function
cmk.notification_plugins.utils.format_plugin_output.

CMK-2142

Change-Id: I865d491957e04a8943e711fcf46e46ced926f081