7895 FIX Download agent output: Fix timeouts in distributed setups
When executing the action "Download agent output" (or SNMP walk) for a host
that is montored by a remote site, this action could result in a timeout when
it takes longer than 120 seconds to fetch the data from the monitored system.

This has now been fixed by moving the background job execution to the remote
site.

Change-Id: I4ecd54c87ba627a2fa60cb2296d86aa06dc94698