7170 FIX cmk-update-agent: Stop stockpiling broken installation packages
After failing to install the downloaded installation package,
the agent updater keeps the package saved at your system's temp-dir
for analysis.

While this may be helpful in some situations, it leads to an extensive
collection of downloaded installation packages if the update fails
repeatedly.

With this werk, the package is not kept any longer after a failed
installation. Instead, you are directed to the matching installation
package located at the checkmk-server.

(CMK-2158)

Change-Id: Ia6d4d1708605bdd63307a45aa6276de9090a8f9f