7431 FIX mk_postgres: version detection may fail
The version detection for the server did not use the
configured database, user and port defined in the
postgres.cfg. Instead the default values were used.
If the plugin cannot execute queries with the default
values the plugin will exit prematurely.

To resolve the issue the fixed version of the mk_postgres
plugin has to be deployed on the affected hosts.

Change-Id: I8e12118c20d96f345c9be8a5f33e9387201ce4a4