7912 FIX Fix host tag update problem
With 1.6 the data structure of tag definitions has been changed
and also moved to another file. WATO is now saving the tags in
<tt>etc/check_mk/conf.d/wato/tags.mk</tt> instead of
<tt>etc/check_mk/conf.d/wato/hosttags.mk</tt>.

There is no manual action needed for migration. During update
from previous versions the old configuration is read and saved
in the new format and file.

In previous 1.6 versions the migration did not correctly delete
the <tt>etc/check_mk/conf.d/wato/hosttags.mk</tt> after saving
it in the new file which is fixed now.

Change-Id: I3638aa2ec3dbb80ab9d2dfaa3bf144090f36cd70