7971 FIX Fixed wrong check interval for "Check_MK HW/SW Inventory" service
Sites created with one of the previous 1.6 releases were using a check interval
of 1 minute instead of the intended 1 day. The problem was that the sample
configuration, which is created during site creation, created a wrong rule in
the ruleset "Normal check interval for service checks". This could be fixed by
creating the rule manually.

Sites created with 1.6.b4 or newer will now use the intended check interval again.

CMK-2371

Change-Id: I73ee2031e51fe07ea15b9b67593af0d6901e6a86