Discovery functions can now produce service labels
Instead of producing a two element tuple of <tt>(item, parameters)</tt>
to create a new servuce the discovery functions can now provide objects
like this:

C+:
Service(
    item="Interface 213",
    parameters=None,
    service_labels=ServiceLabels(
        ServiceLabel(u"check_type", u"network_interface"),
        ServiceLabel(u"interface_type", u"uplink"),
    )
)
C-:

These objects can be returned / yielded like the tuples before. Both,
the parameters and the service_labels are optional arguments.

This feature is not complete yet. In the next steps the autocheck format
needs to be changed to be more flexible. Then the service labels can be
persisted in the autocheck files and read from them for the core config
creation.

CMK-1816

Change-Id: Ie2ce6704a7742f7f6ff21424b3c3bb7b749357b8