Make autocheck format more flexible
The autocheck entry format has now been changed to be dictionary
based. This makes it possible to extend the data associated with a
service in a more flexible way.

The old autocheck format can still be read. New autochecks will be
written in the new format.

In the next steps we add the discovered service labels as new field
to the autochecks.

CMK-1816

Change-Id: I85c6b561d58eb807430e8a167a7468847d95862f