Disentangle local pip2/pip3 installation directories.
Previously we used a .pip/pip.conf file to tell pip where to put new
packages, but this single-file approach doesn't work with a parallel
pip2/pip3 installation anymore.  Therefore, we patch the pip2/pip3
executable wrappers to set the respective environment variable correctly.

As an added tiny bonus, we disable the annoying pip version check: Updating
pip is not a good recommendation for the site user...

Change-Id: I2c70086d290cadd43b28ca369ead02f5c7a0b9f4