Propagate autocheck information a bit further
The read_autochecks_of function, the one which resolves the effective
check parameters for later checking, now returns a list of Service
objects instead of tuples.

These objects contain more information about the service as the tuple
before. It already contains the service description and the service
labels.

This is just the first step. We need to continue with this. All places
that work with CheckTable at the moment need to be changed to use this
structure.

CMK-1816

Change-Id: I1a9d3c885374fa4aef8ea075bf3baa2263efc94a