ListOfMultiple: Choices are now fetched via AJAX on demand
The ListOfMultiple valuespec is currently used in two places (the filter
selection of the view editor and the tag selection of the rule condition
editor).

Instead of precomputing all the possible values, the valuespec is now
only rendering the initially selected choices. In the moment the user
adds a new choice, this will be fetched from the server using an AJAX
call.

Each call site of the ListOfMultiple valuespec needs to implement an
AJAX page which is based on ABCPageListOfMultipleGetChoice.

Change-Id: I42e83b7f009e629f2455b1cbfef18b314f3debc5