Make ruleset match analysis in WATO work - step 1
This is not possible without loading the whole Checkmk base
configuration to get the rulesets, labels and so on.

Using an automation, as we normally do would be a performance overkill.
The component cmk_base now has one module cmk_base.export which exposes
some functionality for other components, like the GUI. This is the same
approach as we use in cmk.ec.export.

We need some more steps to make the analysis really work, like
transporting the service description through several pages of WATO
instead of just the item.

CMK-2402

Change-Id: Iee59a6a3896f4080e227e365b0feaf00ac1da8a2