Make ruleset match analysis in WATO work - step 2
Care about special case of matching on checkgroup parameter rulesets.
These rulesets do not use the whole service_description as condition,
but only the check item name.

This was working transparently until we need the real service description to
gather the labels of a services for matching on rules that use service labels
as conditions. To make this work we now use a specific helper function
cmk_base.config.ruleset_match_object_for_checkgroup_parameters() to get a
ruleset match object that uses the item as service_description and the real
service_description to get service attributes like the service labels.

CMK-2402

Change-Id: I4ab5dea31ffb9b25783d929d9de281aaa6a751e0