8783 FIX bulk notifications did not produce failed notifications
If a notification script returned a non-zero exit code and was
configured for bulk notifications no failed notifications were
shown in the GUI. With this fix one failed notification will
be produced for every notification present in the failed bulk.

Furthermore the view "Host and Service notifications" now shows
a "Final notification result" for every notification in a bulk.

CMK-2079

Change-Id: Ife79e17e6fb734cad8de11555c06d23ced8fb71f