7956 agent_vsphere: Collect performance counters from vCenter
Previously some counters could only be fetched from the ESX host directly.
We now provide the counters when connecting to the vCenter as well.

If the special agent is executed against the vCenter and creating services
on the ESX hosts via the piggyback mechanism, this may add the "Disk IO",
"Datastore IO" and "Interface" services to the affected hosts.

This feature is a result of the fact that we no longer request
the syntax for desired counters, but for all available ones.

This solves CMK-2256.

Change-Id: I335bcff1700e7bca46ade5593473cc89d3f3b704