7960 ps: Option to include process arguments in long-output
This Werk adds an option to include process arguments in the long-output
of the service.
If specified in the ruleset "State and count of processes", the list of
all the matching processes and their details in the long-output will
include up to the first N characters of each processes' arguments.

<b>Note:</b> This feature must be used with care, as the output may
include sensitive data like credentials.

Change-Id: Ib9bca3d5f196e7d8a996c9cdb8c0b5345924a1d5