8847 FIX Service discovery: Print warning instead of failing on discovery errors
The service discovery page of WATO was terminating the service discovery on the
first error, independent of which discovery function raised the error. In this
situation a single broken check could break the service discovery functionality
of WATO.

On the command line (cmk -I) the default behavious is to proceed with the other
discovery function when a single discovery function fails. This is now the
default behavior of the WATO discovery page.

Change-Id: Iec8aa64f1b6d321bcf61de449ea0ec86f0335567