Removed confusing signature change in logger class hierarchy.
Apart from removing a big OO-no-no, this removes one of the two
responsibilities of CMKWebLogger.  The remaining functionality should
probably live in a formatter or filter, but this is a different story.  :-)

As a nice side effect, the exception log messages carry a bit more
information, although there is probably still some room for improvements.

Change-Id: I119f4dc916f03a44fa1dfb206d2ed6065c132bd7