7939 FIX Fixed address => host name mapping for events
When an incoming event console message does not contain a host name, but
only an address (e.g. an SNMP trap), the EC should use the host name
configured for that address.  This functionality was broken, so you e.g.
only saw the IP address of a host in notifications triggered by such
messages. This has been fixed.

Change-Id: Idb47d4973c78d102a27d66a209a0aa202bbdfdde