7964 mk_docker.py: No longer pipe checkmk agent into containers
The plugin mk_docker.py tries to execute the linux agent inside the container.
If the contianer had no agent installed, plugin used to pipe the linux agent
into the docker container. The latter is no longer done.

The agent is still executed if users have installed in on the container.

CMK-2506

Change-Id: I06a006e901b340bf4504c8b1154b11a4adf830d4