Unbreak various checks depending on the cwd and/or previous checks.
The file names given to config.load_checks() should better be absolute,
otherwise loading depends on the current working directory, which is almost
always totally wrong.  We already have a CheckManager class for this, so we
should simply use that.

Note that various test previously "worked" only by accident: The
load_checks() always failed internally, but it simply gobbled up the "file
not found" exception, continuing as if nothing has happened.  If and only if
another test previously loaded the required check, things seemed to work.

Change-Id: Ib65af1c15f48b83bf4c1f29a9da843c75fc8938e