8861 FIX DCD: Gracefully handle single host issues when creating hosts
When one of multiple hosts could not be created by a execution,
this resulted in a complete termination of the execution. A persistant
issue could totally break a connection, since no further steps would
be executed.

As first step the "add_hosts" API call has been extended to report
back the result of a call. The structure looks like this:

C+:
{
    "succeeded_hosts": ["host1", "host3"],
    "failed_hosts": {
        "host2": "an error occured ...",
    },
}
C-:

The DCD reads this response and logs the result. In case there have been hosts
added, the discovery is performed for them. After that, the connection
continues with the following steps.

Change-Id: I10ff0f7af8100e5c545ef335d6657f39ff090265