8863 Discovery page: Add action to update the discovered host labels
The discovery page of a host shows the discovered host labels for the
selected host. It is now possible to persist the new found host labels
using the button "Update host labels" without updating the services.

The "automatic refresh (tabula rasa)" also updates the discovered host
labels in addition to the discovered services of the host.

Change-Id: I97d59274b3f473d9bd7bd1bc1b1eddbf5897a097