Data tables: Only show action row when useful
When having a table that is not searchable, but only sortable and
the user has not yet sorted the table by hand, the action button
row could be toggled but it was empty. After sorting a button
"reset sorting" appeared, which gives sense to the action row.

The row is not shown anymore, when there is no action button
visible in the row.

CMK-2554

Change-Id: Ic644f57be856b68ece2dc9fbdd5ec4802eeb176a