8865 DCD: Add host deletion grace time after site initialization
The hosts created with the DCD are based on the piggyback data produced by a
source host.

This piggyback data is stored in Checkmk sites in a temporary directory below
<tt>tmp/check_mk/piggyback*</tt>. This temporary directory is initialized
during the first startup of a site after the Checkmk host was booted.  From the
initialization to the first contact with the source host, the piggyback data is
missing and could lead to host being removed by the DCD accidentally.

To prevent this, the DCD connections got a new mandatory configuration option
"Prevent host deletion right after initialization". It is used to configure the
duration where no host is deleted after the site (tmpfs) initialization.
It defaults to 10 minutes, which should be enough for the piggyback sources to
produce the needed data.

Change-Id: If7df09d19ecd42c41f9dfb4175a2997b48d8e522