Removed two (local) circular imports.
YAPF and pylint fight against each other a bit, so we have to use some dummy
empty comments to keep the formatting sane.  When we have Python 3, lots of
"unused" imports go away because the typing annotations are visible to
pylint.  Furthermore, more recent pylint versions seem to be less dependent
on the actual line of a pragma.

Change-Id: I54a8dc16a6d2908e85a157be29565c780409df8a