8970 FIX diskstat: fix of parsing functionality concerning device names
This is a bug fix for the diskstat check. This fix primarily concerns users who rely on XEN servers where the name format of discs slightly differs. The updated version makes sure that no discs are accidentally skipper over during the parsing process of the information. The mentioned name format does not concern all users.

Change-Id: Ic93735b9f10ed9b7dbff30eb404efcb8e30d6554