Refactoring: Move funny special cases to the call sites.
Structurally, this is inversion of control, a.k.a. the Hollywood Principle:
"Don't call us, we call you!" If we had classes here at hand (which we
really should), this refactoring would be:

   https://refactoring.com/catalog/replaceConditionalWithPolymorphism.html

We already *have* a MonitoringCore class where the additional three methods
should live, but threading through the right instance is currently a bit
tricky because of several layers of "interesting" code in between.  Without
an instance at hand, we simulate it by passing three functions, i.e. we use
"poor man's OO" for now.  :-)

Change-Id: I245b373220e60ae24977af020c4db928e5486f77