8879 Docker: Enable status data inventory for nodes by default
The factory settings (which are applied only for new sites) are now creating
a rule in the "Do hardware/software inventory" ruleset which enables the status
data inventory for all docker nodes (hosts with label "cmk/docker_object:node").

This simplifies the initial docker monitoring setup.

Users which update existing sites need to create this rule on their own.

Change-Id: I93561ee7034cbf75445837963aee02a27a417c5c