postgres: Bugfix in postgres_version and posgres_stat_database for instance down
The check postgres_version creates a CRIT when a PostgreSQL Cluster is not
running. This issue is a regression of Werk 7140 which introduced this issue.

The check postgres_stat_database creates a CRIT when a PostgreSQL Cluster is not
running. This is useless for a check with statistical informations only.

Both checks goes PEND when agent output is missing.