8882 FIX Fix invalid rule match analyzation in WATO
The "Parameters of host/service" page and ruleset lists showed an invalid match
result for rules which value evaluates to None.  For binary rulesets (the ones
that only have yes/no as value) the result was also not displayed correctly.

The ruleset matching of Checkmk during checking was not affected by this issue.

Change-Id: I1c9f4b2747332d09ca40deac85ecfc0fe9fce2b9