10056 Windows Agent command line parameter 'reload_config' starts/stops plugins if required
Earlier 'reload_config' did not try to run asynchronous plugins.

Thus, the parameter was useless for the automatic update of the
agent with the Agent Updater plugin.

Since now, the Windows Agent starts and stops the asynchronous plugins according
to the configuration just loaded with command line parameter 'reload_config'.

Change-Id: I464773a4ac035fd0b606d5447c7344f1f136ecb2