8911 FIX Fix CRIT criterion for k8s_pod_container
Cleanly terminated containers would erroneously mark the
check CRIT.  Now, the check is only CRIT when a container
exited with a non-zero error code, otherwise it remains OK.
The check now reports more information: the number of
running containers, the number of waiting containers, and
the number of terminated containers.

This fixes CMK-2683.

Change-Id: I8b76b5e71f1d5ca2f4539407a64735bde298d789