10038 FIX Fix cmk/check_mk_server:yes label related issues
This label was detected based on the asynchronously created agent section
'<<<omd_status>>>'. Because this section is not immediately available on the
first agent contact, this resulted in a label discovery after the second check
cycle.

Another issue was that the '<<<omd_status>>>' section was produced by non
Checkmk servers with older agent versions. This lead to a wrong label discovery.

We now use the dedicated '<<<labels>>>' section to create the label which is
produced once we find the command 'omd' on a system.

Change-Id: I81bf19060b4d7c8f98efd28472cf4855508a4ce9