10040 FIX Fix wrong core host name in events when using host name rewriting
The Event Console queries the monitoring core to find a match between the host
in the core and the origin of the event. In case you used the event rewrite
mechanism to change the host name, the lookup in the core was done with the
original host name of the event instead of the rewritten host name.

CMK-2702

Change-Id: Ide37e44605a61846493c3dab8cd86053d862d176