Improve Python 2/3 compatible Path import
The previous import based on try/expect did not work as intended,
because mypy did not understood that pathlib is only available with
Python 3. To suppress the import issues we added "# type: ignore" to
that line, which implicitly marked Path with the "Any" type. This way
it was not possible for mypy to detect issues with Path.

This new approach uses the recommended way of python version checks (see
https://github.com/python/mypy/blob/master/docs/source/common_issues.rst#python-version-and-system-platform-checks).

With this variant it is now possible to detect wrong operations on
the Path object in our code which is meant to be portable during
the 2/3 transition (cmk.utils).

Change-Id: Ie37da4102c39e605cd0bfc5a48a2fc70f443109a